% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{convert_state_abb_to_name}
\alias{convert_state_abb_to_name}
\title{Convert state abbreviation to name}
\usage{
convert_state_abb_to_name(state = NULL)
}
\arguments{
\item{state}{character for state name. May either be an abbreviation or full name and
may be either lower-case or upper-case.}
}
\value{
lower-case character representing full state name.
}
\description{
Convert U.S. state abbreviation to full name.
}
\details{
Used by other package functions to coerce \code{state} parameter to proper format
for working with \code{ggplot2::map_data()}
}
