% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{create_map_state}
\alias{create_map_state}
\title{Get data for a \code{ggplot2} map of a single U.S. state}
\usage{
create_map_state(state = NULL, show_county = TRUE, fill = "white",
  color = "black")
}
\arguments{
\item{state}{character for state name. May either be an abbreviation or full name and
may be either lower-case or upper-case.}

\item{show_county}{logical. Whether or not to show county lines for specified state.}

\item{fill, color}{character. Either a hex code or a valid color name.
Parameters passed directly to \code{ggplot2::geom_polygon()}.}
}
\value{
gg.
}
\description{
Create the state layer for a \code{ggplot2} plot.
}
\details{
Only works for a single state.
Note that when extra layers are added, they must specify a \code{group} (e.g. \code{group = ""}),
in the \code{ggplot2::aes()} call. Otherwise, an error message is returned.
}
\seealso{
\code{ggplot2::map_data()}
}
