% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_color_inv}
\alias{get_color_inv}
\title{Get the inverse of a color}
\usage{
get_color_inv(color = NULL)
}
\arguments{
\item{color}{character. Either a hex code or a valid color name.}
}
\value{
character.
}
\description{
None.
}
\details{
used by the \code{create_map_state()} function.
}
