% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{get_map_data_county}
\alias{get_map_data_county}
\alias{get_map_data_county_tx}
\alias{get_map_data_state}
\alias{get_map_data_state_tx}
\title{Get data for \code{ggplot2} U.S. state map}
\usage{
get_map_data_county(state = NULL)

get_map_data_county_tx(state = "texas")

get_map_data_state(state = NULL)

get_map_data_state_tx(state = "texas")
}
\arguments{
\item{state}{character for state name. May either be an abbreviation or full name and
may be either lower-case or upper-case.}
}
\value{
data.frame.
}
\description{
Work with \code{ggplot2::map_date()}.
}
\details{
Note that \code{region} is the column name for states,
and \code{subregion} is the column name for counties.
(This is the convention used by \code{ggplot2::map_date()}.
}
\seealso{
\code{ggplot2::map_data()}
}
