% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggmap_stamen_tx}
\alias{ggmap_stamen_tx}
\title{Texas map for \code{ggplot2} maps}
\format{gg}
\usage{
ggmap_stamen_tx
}
\description{
\code{gg} base layer for state of Texas.
}
\details{
Transformed version of raw ggmap object without axis text or titles.
}
\keyword{datasets}
