% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ggmap_stamen_tx_raw}
\alias{ggmap_stamen_tx_raw}
\title{Texas map for \code{ggplot2} maps}
\format{ggmap}
\usage{
ggmap_stamen_tx_raw
}
\description{
\code{ggmap} base layer for state of Texas.
}
\details{
Raw download using \code{ggmap} package.
}
\keyword{datasets}
