% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_te.R
\name{scale_color_te}
\alias{scale_color_te}
\title{Color scale constructor for \code{te_color}}
\source{
\url{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}.
\url{https://github.com/hrbrmstr/hrbrthemes/blob/master/R/scales.r}
}
\usage{
scale_color_te(palette = "discrete", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{character. Name of palette in \code{te_palettes} list.}

\item{discrete}{logical. Indicates whether the color palette is discrete or not.}

\item{reverse}{logical. Indicates whether the palette should be reversed.}

\item{...}{dots. Additional arameters passed to \code{ggplot2::discrete_scale()} if
\code{discrete = TRUE} or
\code{ggplot2::scale_fill_gradientn()} if \code{discrete = FALSE}.}
}
\description{
Color function for use with \code{ggplot2}.
}
\details{
None.
}
