% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_y_pretty_comma}
\alias{scale_y_pretty_comma}
\alias{scale_y_pretty_percent}
\alias{scale_x_pretty_comma}
\alias{scale_x_pretty_percent}
\alias{scale_xy_pretty_comma}
\alias{scale_xy_pretty_percent}
\title{Create a scale to add to a \code{ggplot2} plot}
\usage{
scale_y_pretty_comma(...)

scale_y_pretty_percent(...)

scale_x_pretty_comma(...)

scale_x_pretty_percent(...)

scale_xy_pretty_comma(...)

scale_xy_pretty_percent(...)
}
\arguments{
\item{...}{dots. Arguments passed to underlying \code{ggplot2} function.}
}
\value{
gg object.
}
\description{
Creates pretty breaks for a continuous scale.
}
\details{
Only works for a single state.
}
