% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_color_set1}
\alias{scale_color_set1}
\alias{scale_fill_set1}
\title{Quick access to RColorBrewer palette 'Set1'.}
\usage{
scale_color_set1()

scale_fill_set1()
}
\value{
gg object.
}
\description{
Use to incorporate RColorBrewer palette 'Set1' in \code{ggplot}s.
}
\details{
This is a color palette that I use quite often (simply
as an alternative to \code{ggplot2}'s default palette.
}
