% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spdf_tx}
\alias{spdf_tx}
\title{Texas map for \code{ggplot2} maps}
\format{SpatialPolygonsDataFrame}
\source{
\url{http://colby.edu/~mgimond/Spatial/Data//texas.rds}
}
\usage{
spdf_tx
}
\description{
SpatialPolygonsDataFrame layer for state of Texas.
}
\details{
Converted to \code{sp::CRS("+proj=longlat +datum=WGS84")}.
}
\seealso{
\url{https://mgimond.github.io/Spatial/interpolation-in-r.html}
}
\keyword{datasets}
