% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spdf_tx_precip}
\alias{spdf_tx_precip}
\title{Data for \code{spdf_tx} map example}
\format{SpatialPointsDataFrame}
\source{
\url{http://colby.edu/~mgimond/Spatial/Data/pricip.rds}
}
\usage{
spdf_tx_precip
}
\description{
Data to be used for examples with SpatialPolygonsDataFrame class data.
}
\details{
No transformations applied to the raw data.
}
\seealso{
\url{https://mgimond.github.io/Spatial/interpolation-in-r.html}
}
\keyword{datasets}
