% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_te.R
\docType{data}
\name{te_colors}
\alias{te_colors}
\title{Custom colors}
\format{character (vector). Named.}
\usage{
te_colors
}
\description{
colors
}
\details{
These were hand-picked, keeping in mind 'oppositional' colors
(i.e. red and green, blue and orange, purple and yellow, etc.)
}
\keyword{datasets}
