% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_te.R
\name{te_cols}
\alias{te_cols}
\title{Extract hex values from te_colors}
\usage{
te_cols(...)
}
\arguments{
\item{...}{dots. Names of \code{te_colors} to subset.}
}
\description{
Extract hex values from te_colors
}
