% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_te}
\alias{theme_te}
\alias{theme_te_dx}
\alias{theme_te_facet}
\alias{theme_te_facet_dx}
\title{Custom theme}
\usage{
theme_te(..., option = c("a", "b"))

theme_te_dx(option = c("a", "b"), ...)

theme_te_facet(..., option = c("a", "b"))

theme_te_facet_dx(..., option = c("a", "b"))
}
\arguments{
\item{...}{dots. Parameters passed to theme specified by option.}

\item{option}{character. Either "a" or "b".}
}
\description{
A custom \code{ggplot2} theme.
}
\details{
Option "a" is an 'original' theme.
Option "b" is an adapted \code{hrbrthemes::theme_ipusm()} theme.
}
