% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_te_a}
\alias{theme_te_a}
\alias{theme_te_a_dx}
\alias{theme_te_a_facet}
\alias{theme_te_a_facet_dx}
\title{Custom theme}
\usage{
theme_te_a(base_family = "", base_size = 11, plot_title_size = 16,
  subtitle_size = 12, ...)

theme_te_a_dx(...)

theme_te_a_facet(...)

theme_te_a_facet_dx(...)
}
\arguments{
\item{base_family}{character.}

\item{base_size, plot_title_size, subtitle_size}{numeric.}

\item{...}{dots. Additional parameters passed to \code{ggplot2::theme()}.}
}
\description{
A custom \code{ggplot2} theme.
}
\details{
None.
}
\seealso{
\url{https://github.com/hrbrmstr/hrbrthemes/blob/master/R/theme-ipsum.r}
}
