% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_te_tile}
\alias{theme_te_tile}
\title{Custom theme for tile plot}
\usage{
theme_te_tile(base_family = "Arial Narrow", base_size = 12,
  panel.grid.major = ggplot2::element_blank(),
  panel.grid.minor = ggplot2::element_blank(),
  axis.text.x = ggplot2::element_text(angle = 90), ...)
}
\arguments{
\item{base_family}{character.}

\item{base_size}{numeric.}

\item{panel.grid.major, panel.grid.minor, axis.text.x}{gg elements.
Set to a non-trivial default
Passed directly to \code{ggplot2::theme()} parameters with similar names.}

\item{...}{dots. Passed directly to \code{ggplot2::theme_bw()}}
}
\description{
A custom \code{ggplot2} theme.
}
\details{
Uses \code{theme_te()} as basis.
}
