% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tmap_tx}
\alias{tmap_tx}
\title{Texas map for \code{ggplot2} maps}
\format{tmap}
\usage{
tmap_tx
}
\description{
tmap layer for state of Texas.
}
\details{
Transformed version of \code{spdf_tx}, where \code{tmap::tm_polygons(col = "white", alpha = 0)}
has been called after conversion to a tmap object (via \code{tmap::tm_shape(spdf_tx)}).
}
\seealso{
\url{https://mgimond.github.io/Spatial/interpolation-in-r.html}
}
\keyword{datasets}
