% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{update_geom_font_defaults}
\alias{update_geom_font_defaults}
\title{Update \code{ggplot2} font defaults for text geoms}
\source{
\url{https://github.com/hrbrmstr/hrbrthemes/blob/master/R/theme-ipsum.r}.
}
\usage{
update_geom_font_defaults(family = "Arial Narrow", face = "plain",
  size = 3.5, color = "#2b2b2b")
}
\arguments{
\item{family, face, size, color}{character.}
}
\description{
Updates \code{ggplot2::geom_label()} and \code{ggplot2::geom_text()} font defaults.
}
\details{
None.
}
