% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_te.R
\name{te_pal}
\alias{te_pal}
\title{Return function to interpolate a te_color palette}
\source{
\url{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}.
\url{https://github.com/hrbrmstr/hrbrthemes/blob/master/R/scales.r}
}
\usage{
te_pal(palette = "discrete", discrete = TRUE, reverse = FALSE)
}
\arguments{
\item{palette}{character. Name of palette in \code{te_palettes} list.}

\item{discrete}{logical. Indicates whether the color palette is discrete or not.}

\item{reverse}{logical. Indicates whether the palette should be reversed.}
}
\description{
Returns a function to be used by other functions.
}
\details{
Calls \code{scales::manual_pal()} if \code{discrete = TRUE} or  \code{colorRampPalette()} if
\code{discrete = FALSE}.
}
