% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_te}
\alias{theme_te}
\alias{theme_te_dx}
\alias{theme_te_facet}
\alias{theme_te_facet_dx}
\title{Custom theme}
\usage{
theme_te(base_family = "Arial Narrow", base_size = 12, title.size = 16,
  title.face = "bold", subtitle.size = base_size, subtitle.face = "plain",
  caption.size = base_size, caption.face = subtitle.face,
  strip.text.size = base_size, strip.text.face = subtitle.face,
  axis.title.size = base_size, axis.title.face = subtitle.face,
  axis.text.size = base_size, axis.text.face = subtitle.face, ...)

theme_te_dx(...)

theme_te_facet(...)

theme_te_facet_dx(...)
}
\arguments{
\item{base_family, title.face, subtitle.face, caption.face, strip.text.face, axis.title.face, axis.text.face}{character.}

\item{base_size, title.size, subtitle.size, caption.size, strip.text.size, axis.title.size, axis.text.size}{numeric.}

\item{...}{dots. Additional parameters passed to \code{ggplot2::theme()}.}
}
\description{
A custom \code{ggplot2} theme.
}
\details{
Uses \code{ggplot2::theme_minimal()} as basis.
}
\seealso{
\url{https://github.com/hrbrmstr/hrbrthemes/blob/master/R/theme-ipsum.r}
}
