% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_te}
\alias{theme_te}
\alias{theme_te_dx}
\alias{theme_te_facet}
\alias{theme_te_facet_dx}
\title{Custom theme}
\usage{
theme_te(base_family = "Arial Narrow", base_size = 12, title.size = 16,
  title.face = "bold", subtitle.size = base_size, subtitle.face = "plain",
  caption.size = base_size, caption.face = subtitle.face,
  strip.text.size = base_size, strip.text.face = subtitle.face,
  axis.title.size = base_size, axis.title.face = subtitle.face,
  axis.text.size = base_size, axis.text.face = subtitle.face,
  legend.position = "bottom", legend.title = ggplot2::element_blank(),
  panel.grid.minor = ggplot2::element_blank(),
  plot.background = ggplot2::element_blank(), panel.spacing = grid::unit(1,
  "lines"), axis.title.just = "rt", ...)

theme_te_dx(...)

theme_te_facet(...)

theme_te_facet_dx(...)
}
\arguments{
\item{base_family, title.face, subtitle.face, caption.face, strip.text.face, axis.title.face, axis.text.face}{character.}

\item{base_size, title.size, subtitle.size, caption.size, strip.text.size, axis.title.size, axis.text.size}{numeric.}

\item{legend.position}{gg element. Set to \code{"bottom"} by default
(instead of "right").}

\item{legend.title}{gg element. Set to a non-trivial default
(instead of \code{ggplot2::element_text()}).}

\item{panel.grid.minor}{gg element. Set to a non-trivial default
(instead of \code{ggplot2::element_line()}).}

\item{plot.background}{gg element. Was set to a non-trivial default
(instead of \code{ggplot2::element_blank()}) in a previous version.}

\item{panel.spacing}{gg element. Set to a non-trivial default
(instead of \code{5.5pt}).}

\item{axis.title.just}{character. One of one of \code{'blmcrt'}}

\item{...}{dots. Additional parameters passed to \code{ggplot2::theme()}.}
}
\description{
A custom \code{ggplot2} theme.
}
\details{
Uses \code{ggplot2::theme_minimal()} as basis. Customized
for personal preferences.
}
\seealso{
\url{https://github.com/hrbrmstr/hrbrthemes/blob/master/R/theme-ipsum.r}.
\url{\code{hrbrthemes::theme_ipsum()}}
}
