% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_te_b}
\alias{theme_te_b}
\alias{theme_te_b_dx}
\alias{theme_te_b_facet}
\alias{theme_te_b_facet_dx}
\title{Custom theme (inspired by hrbrmstr)}
\usage{
theme_te_b(base_family = "Arial Narrow", base_size = 12,
  title.family = base_family, title.size = 18, title.face = "bold",
  title.margin = 10, subtitle.family = base_family, subtitle.size = 12,
  subtitle.face = "plain", subtitle.margin = 10,
  strip.text.family = base_family, strip.text.size = 12,
  strip.text.face = "plain", caption.family = base_family,
  caption.size = 12, caption.face = "plain", caption.margin = 10,
  axis.text.size = base_size, axis.title.family = subtitle.family,
  axis.title.size = 12, axis.title.face = "plain", axis.title.just = "rt",
  plot.margin = ggplot2::margin(1, 1, 1, 1), ...)

theme_te_b_dx(...)

theme_te_b_facet(...)

theme_te_b_facet_dx(...)
}
\arguments{
\item{base_family, title.face, subtitle.face, strip.text.face, caption.face, axis.title.face}{character.}

\item{base_size, title.size, title.margin, subtitle.size, subtitle.margin, strip.text.size, caption.size, caption.margin, axis.title.size, axis.text.size}{numeric.}

\item{title.family, subtitle.family, strip.text.family, caption.family, axis.title.family}{character. Inherit from base_family.}

\item{axis.title.just}{character.}

\item{plot.margin}{numeric vector.}

\item{...}{dots. Additional parameters passed to \code{ggplot2::theme()}.}
}
\description{
Custom \code{ggplot2} theme that is based on \code{hrbrthemes::theme_ipsum()}.
}
\details{
Uses \code{ggplot2::theme_minimal()} as basis.
Legend is formatted in a different manner than \code{hrbrthemes::theme_ipsum()}.
Parameters not borrowed: \code{grid_col, grid, axis.co, axis, ticks}.
}
