% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_map}
\alias{theme_map}
\title{Custom theme for maps}
\usage{
theme_map(base_family = "Arial Narrow", base_size = 12,
  panel.background.fill = NULL, panel.background.color = NULL,
  strip.background.fill = panel.background.fill,
  strip.background.color = panel.background.color, ...)
}
\arguments{
\item{base_family}{character.}

\item{base_size}{numeric.}

\item{panel.background.fill, panel.background.color, strip.background.fill, strip.background.color}{character.
Passed directly to \code{ggplot2::theme()} parameters with similar names.}

\item{...}{dots. Passed directly to \code{theme_te()}}
}
\description{
A custom \code{ggplot2} theme.
}
\details{
Uses \code{theme_te()} as basis.
}
\seealso{
\url{http://eriqande.github.io/rep-res-web/lectures/making-maps-with-R.html}.
\url{https://gist.github.com/hrbrmstr/33baa3a79c5cfef0f6df}.
}
